[main] restructure
