+++ update.go
@@ -1,0 +1,111 @@
+package main
+
+import (
+	"github.com/charmbracelet/bubbles/key"
+	"github.com/charmbracelet/bubbles/viewport"
+	tea "github.com/charmbracelet/bubbletea"
+	"github.com/charmbracelet/lipgloss"
+	"github.com/dbd/ght/components"
+	"github.com/dbd/ght/components/pullRequestDetail"
+	"github.com/dbd/ght/components/pullRequestSearch"
+	"github.com/dbd/ght/components/tab"
+)
+
+func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
+	var cmds []tea.Cmd
+	activeTab := m.Tabs[m.activeTab]
+	switch msg := msg.(type) {
+	case pullRequestSearch.OpenPR:
+		var alreadyOpened bool
+		for counter, tab := range m.Tabs {
+			if tab.Name == msg.PR.Title {
+				alreadyOpened = true
+				_ = counter
+			}
+		}
+		if alreadyOpened {
+			m.activeTab = len(m.Tabs) - 1
+			m.focused = true
+		} else {
+			t := tab.NewModel(m.context, msg.PR.Title)
+			t.Page = pullRequestDetail.NewModel(msg.PR, m.context)
+			m.Tabs = append(m.Tabs, t)
+			m.activeTab = len(m.Tabs) - 1
+			m.focused = true
+		}
+	case components.Blur:
+		m.focused = true
+	case tea.WindowSizeMsg:
+		headerHeight := lipgloss.Height(m.headerView())
+		footerHeight := lipgloss.Height(m.footerView())
+		verticalMarginHeight := headerHeight + footerHeight
+
+		if !m.ready {
+			m.viewport = viewport.New(msg.Width, msg.Height-verticalMarginHeight)
+			m.viewport.YPosition = headerHeight - 1
+			m.context.ViewportWidth = m.viewport.Width
+			m.context.ViewportHeight = m.viewport.Height
+			m.context.ViewportYOffset = m.viewport.YOffset
+			m.context.ViewportYPosition = m.viewport.YPosition
+			m.viewport.SetContent(activeTab.Page.View())
+			_, cmd := activeTab.Update(msg)
+			cmds = append(cmds, cmd)
+			m.ready = true
+		} else {
+			m.viewport.Width = msg.Width
+			m.viewport.Height = msg.Height - verticalMarginHeight
+			m.context.ViewportWidth = m.viewport.Width
+			m.context.ViewportHeight = m.viewport.Height
+			m.context.ViewportYOffset = m.viewport.YOffset
+			m.context.ViewportYPosition = m.viewport.YPosition
+			activeTab, cmd := activeTab.Update(msg)
+			cmds = append(cmds, cmd)
+			m.Tabs[m.activeTab] = activeTab
+		}
+	case tea.KeyMsg:
+		if !m.focused {
+			activeTab, cmd := activeTab.Update(msg)
+			cmds = append(cmds, cmd)
+			m.Tabs[m.activeTab] = activeTab
+		} else {
+			switch {
+			case key.Matches(msg, m.context.KeyMap.Help):
+				m.showHelp = !m.showHelp
+			case key.Matches(msg, m.context.KeyMap.Down):
+				m.focused = false
+			case key.Matches(msg, m.context.KeyMap.Exit):
+				return m, tea.Quit
+			case key.Matches(msg, m.context.KeyMap.Close):
+				var tt []tab.Model
+				for counter, tab := range m.Tabs {
+					if counter != m.activeTab {
+						tt = append(tt, tab)
+					}
+				}
+				if len(tt) == 0 {
+					m.context.StatusText = "Unable to close last tab. Exit instead."
+					break
+				}
+				m.Tabs = tt
+				if m.activeTab < len(m.Tabs)-1 {
+					m.activeTab++
+				} else if m.activeTab != 0 {
+					m.activeTab--
+				}
+			case key.Matches(msg, m.context.KeyMap.Left):
+				if m.activeTab > 0 {
+					m.activeTab--
+				}
+			case key.Matches(msg, m.context.KeyMap.Right):
+				if m.activeTab < len(m.Tabs)-1 {
+					m.activeTab++
+				}
+			}
+		}
+	default:
+		activeTab, cmd := activeTab.Update(msg)
+		cmds = append(cmds, cmd)
+		m.Tabs[m.activeTab] = activeTab
+	}
+	return m, tea.Batch(cmds...)
+}
